# Replace uses of `std::shared_ptr` with shared ptr without atomic refcounts