--- a/hilti/runtime/include/3rdparty/boost/aligned_storage.hpp
+++ b/hilti/runtime/include/3rdparty/boost/aligned_storage.hpp
@@ -1,0 +1,1 @@
+../../../../../3rdparty/boost/type_traits/include/boost/aligned_storage.hpp