--- a/hilti/runtime/include/3rdparty/boost/assert
+++ b/hilti/runtime/include/3rdparty/boost/assert
@@ -1,0 +1,1 @@
+../../../../../3rdparty/boost/assert/include/boost/assert