--- a/hilti/runtime/include/3rdparty/boost/assert.hpp
+++ b/hilti/runtime/include/3rdparty/boost/assert.hpp
@@ -1,0 +1,1 @@
+../../../../../3rdparty/boost/assert/include/boost/assert.hpp