--- a/hilti/runtime/include/3rdparty/boost/config
+++ b/hilti/runtime/include/3rdparty/boost/config
@@ -1,0 +1,1 @@
+../../../../../3rdparty/boost/config/include/boost/config