--- a/hilti/runtime/include/3rdparty/boost/config.hpp
+++ b/hilti/runtime/include/3rdparty/boost/config.hpp
@@ -1,0 +1,1 @@
+../../../../../3rdparty/boost/config/include/boost/config.hpp