--- a/hilti/runtime/include/3rdparty/boost/core
+++ b/hilti/runtime/include/3rdparty/boost/core
@@ -1,0 +1,1 @@
+../../../../../3rdparty/boost/core/include/boost/core