--- a/hilti/runtime/include/3rdparty/boost/cstdint.hpp
+++ b/hilti/runtime/include/3rdparty/boost/cstdint.hpp
@@ -1,0 +1,1 @@
+../../../../../3rdparty/boost/config/include/boost/cstdint.hpp