--- a/hilti/runtime/include/3rdparty/boost/current_function.hpp
+++ b/hilti/runtime/include/3rdparty/boost/current_function.hpp
@@ -1,0 +1,1 @@
+../../../../../3rdparty/boost/assert/include/boost/current_function.hpp