--- a/hilti/runtime/include/3rdparty/boost/cxx11_char_types.hpp
+++ b/hilti/runtime/include/3rdparty/boost/cxx11_char_types.hpp
@@ -1,0 +1,1 @@
+../../../../../3rdparty/boost/config/include/boost/cxx11_char_types.hpp