--- a/hilti/runtime/include/3rdparty/boost/detail/atomic_count.hpp
+++ b/hilti/runtime/include/3rdparty/boost/detail/atomic_count.hpp
@@ -1,0 +1,1 @@
+../../../../../../3rdparty/boost/smart_ptr/include/boost/detail/atomic_count.hpp