--- a/hilti/runtime/include/3rdparty/boost/detail/iterator.hpp
+++ b/hilti/runtime/include/3rdparty/boost/detail/iterator.hpp
@@ -1,0 +1,1 @@
+../../../../../../3rdparty/boost/core/include/boost/detail/iterator.hpp