--- a/hilti/runtime/include/3rdparty/boost/detail/lightweight_mutex.hpp
+++ b/hilti/runtime/include/3rdparty/boost/detail/lightweight_mutex.hpp
@@ -1,0 +1,1 @@
+../../../../../../3rdparty/boost/smart_ptr/include/boost/detail/lightweight_mutex.hpp