--- a/hilti/runtime/include/3rdparty/boost/detail/lightweight_thread.hpp
+++ b/hilti/runtime/include/3rdparty/boost/detail/lightweight_thread.hpp
@@ -1,0 +1,1 @@
+../../../../../../3rdparty/boost/smart_ptr/include/boost/detail/lightweight_thread.hpp