--- a/hilti/runtime/include/3rdparty/boost/detail/no_exceptions_support.hpp
+++ b/hilti/runtime/include/3rdparty/boost/detail/no_exceptions_support.hpp
@@ -1,0 +1,1 @@
+../../../../../../3rdparty/boost/core/include/boost/detail/no_exceptions_support.hpp