--- a/hilti/runtime/include/3rdparty/boost/detail/quick_allocator.hpp
+++ b/hilti/runtime/include/3rdparty/boost/detail/quick_allocator.hpp
@@ -1,0 +1,1 @@
+../../../../../../3rdparty/boost/smart_ptr/include/boost/detail/quick_allocator.hpp