--- a/hilti/runtime/include/3rdparty/boost/detail/scoped_enum_emulation.hpp
+++ b/hilti/runtime/include/3rdparty/boost/detail/scoped_enum_emulation.hpp
@@ -1,0 +1,1 @@
+../../../../../../3rdparty/boost/core/include/boost/detail/scoped_enum_emulation.hpp