--- a/hilti/runtime/include/3rdparty/boost/detail/sp_typeinfo.hpp
+++ b/hilti/runtime/include/3rdparty/boost/detail/sp_typeinfo.hpp
@@ -1,0 +1,1 @@
+../../../../../../3rdparty/boost/core/include/boost/detail/sp_typeinfo.hpp