--- a/hilti/runtime/include/3rdparty/boost/detail/workaround.hpp
+++ b/hilti/runtime/include/3rdparty/boost/detail/workaround.hpp
@@ -1,0 +1,1 @@
+../../../../../../3rdparty/boost/config/include/boost/detail/workaround.hpp