--- a/hilti/runtime/include/3rdparty/boost/enable_shared_from_this.hpp
+++ b/hilti/runtime/include/3rdparty/boost/enable_shared_from_this.hpp
@@ -1,0 +1,1 @@
+../../../../../3rdparty/boost/smart_ptr/include/boost/enable_shared_from_this.hpp