--- a/hilti/runtime/include/3rdparty/boost/exception
+++ b/hilti/runtime/include/3rdparty/boost/exception
@@ -1,0 +1,1 @@
+../../../../../3rdparty/boost/throw_exception/include/boost/exception