--- a/hilti/runtime/include/3rdparty/boost/get_pointer.hpp
+++ b/hilti/runtime/include/3rdparty/boost/get_pointer.hpp
@@ -1,0 +1,1 @@
+../../../../../3rdparty/boost/core/include/boost/get_pointer.hpp