--- a/hilti/runtime/include/3rdparty/boost/iterator.hpp
+++ b/hilti/runtime/include/3rdparty/boost/iterator.hpp
@@ -1,0 +1,1 @@
+../../../../../3rdparty/boost/core/include/boost/iterator.hpp