--- a/hilti/runtime/include/3rdparty/boost/make_unique.hpp
+++ b/hilti/runtime/include/3rdparty/boost/make_unique.hpp
@@ -1,0 +1,1 @@
+../../../../../3rdparty/boost/smart_ptr/include/boost/make_unique.hpp