--- a/hilti/runtime/include/3rdparty/boost/pointer_cast.hpp
+++ b/hilti/runtime/include/3rdparty/boost/pointer_cast.hpp
@@ -1,0 +1,1 @@
+../../../../../3rdparty/boost/smart_ptr/include/boost/pointer_cast.hpp