--- a/hilti/runtime/include/3rdparty/boost/pointer_to_other.hpp
+++ b/hilti/runtime/include/3rdparty/boost/pointer_to_other.hpp
@@ -1,0 +1,1 @@
+../../../../../3rdparty/boost/smart_ptr/include/boost/pointer_to_other.hpp