--- a/hilti/runtime/include/3rdparty/boost/shared_array.hpp
+++ b/hilti/runtime/include/3rdparty/boost/shared_array.hpp
@@ -1,0 +1,1 @@
+../../../../../3rdparty/boost/smart_ptr/include/boost/shared_array.hpp