--- a/hilti/runtime/include/3rdparty/boost/shared_ptr.hpp
+++ b/hilti/runtime/include/3rdparty/boost/shared_ptr.hpp
@@ -1,0 +1,1 @@
+../../../../../3rdparty/boost/smart_ptr/include/boost/shared_ptr.hpp