--- a/hilti/runtime/include/3rdparty/boost/smart_ptr
+++ b/hilti/runtime/include/3rdparty/boost/smart_ptr
@@ -1,0 +1,1 @@
+../../../../../3rdparty/boost/smart_ptr/include/boost/smart_ptr