--- a/hilti/runtime/include/3rdparty/boost/static_assert.hpp
+++ b/hilti/runtime/include/3rdparty/boost/static_assert.hpp
@@ -1,0 +1,1 @@
+../../../../../3rdparty/boost/static_assert/include/boost/static_assert.hpp