--- a/hilti/runtime/include/3rdparty/boost/swap.hpp
+++ b/hilti/runtime/include/3rdparty/boost/swap.hpp
@@ -1,0 +1,1 @@
+../../../../../3rdparty/boost/core/include/boost/swap.hpp