--- a/hilti/runtime/include/3rdparty/boost/throw_exception.hpp
+++ b/hilti/runtime/include/3rdparty/boost/throw_exception.hpp
@@ -1,0 +1,1 @@
+../../../../../3rdparty/boost/throw_exception/include/boost/throw_exception.hpp