--- a/hilti/runtime/include/3rdparty/boost/type.hpp
+++ b/hilti/runtime/include/3rdparty/boost/type.hpp
@@ -1,0 +1,1 @@
+../../../../../3rdparty/boost/core/include/boost/type.hpp