--- a/hilti/runtime/include/3rdparty/boost/type_traits
+++ b/hilti/runtime/include/3rdparty/boost/type_traits
@@ -1,0 +1,1 @@
+../../../../../3rdparty/boost/type_traits/include/boost/type_traits