--- a/hilti/runtime/include/3rdparty/boost/type_traits.hpp
+++ b/hilti/runtime/include/3rdparty/boost/type_traits.hpp
@@ -1,0 +1,1 @@
+../../../../../3rdparty/boost/type_traits/include/boost/type_traits.hpp