--- a/hilti/runtime/include/3rdparty/boost/utility/addressof.hpp
+++ b/hilti/runtime/include/3rdparty/boost/utility/addressof.hpp
@@ -1,0 +1,1 @@
+../../../../../../3rdparty/boost/core/include/boost/utility/addressof.hpp