--- a/hilti/runtime/include/3rdparty/boost/utility/declval.hpp
+++ b/hilti/runtime/include/3rdparty/boost/utility/declval.hpp
@@ -1,0 +1,1 @@
+../../../../../../3rdparty/boost/type_traits/include/boost/utility/declval.hpp