--- a/hilti/runtime/include/3rdparty/boost/utility/explicit_operator_bool.hpp
+++ b/hilti/runtime/include/3rdparty/boost/utility/explicit_operator_bool.hpp
@@ -1,0 +1,1 @@
+../../../../../../3rdparty/boost/core/include/boost/utility/explicit_operator_bool.hpp