--- a/hilti/runtime/include/3rdparty/boost/version.hpp
+++ b/hilti/runtime/include/3rdparty/boost/version.hpp
@@ -1,0 +1,1 @@
+../../../../../3rdparty/boost/config/include/boost/version.hpp