--- a/hilti/runtime/include/3rdparty/boost/weak_ptr.hpp
+++ b/hilti/runtime/include/3rdparty/boost/weak_ptr.hpp
@@ -1,0 +1,1 @@
+../../../../../3rdparty/boost/smart_ptr/include/boost/weak_ptr.hpp