--- a/hilti/runtime/include/types/reference.h
+++ b/hilti/runtime/include/types/reference.h
@@ -11,6 +11,7 @@
 #include <hilti/rt/any.h>
 #include <hilti/rt/extension-points.h>
 #include <hilti/rt/types/bytes.h>
+#include <hilti/rt/types/shared_ptr.h>
 #include <hilti/rt/types/string.h>
 #include <hilti/rt/types/struct.h>
 #include <hilti/rt/util.h>
@@ -23,7 +24,7 @@ void __attribute__((noreturn)) throw_null();
 
 /** Base for classes that `ValueReference::self` can receive.  */
 template<typename T>
-using Controllable = std::enable_shared_from_this<T>;
+using Controllable = EnableSharedFromThis<T>;
 
 /**
  * Class representing HILTI's `value_ref<T>` type. This class stores an value
@@ -56,7 +57,7 @@ public:
      * Instantiates a reference containing a new value of `T` initialized to
      * its default value.
      */
-    ValueReference() : _ptr(std::make_shared<T>()) {}
+    ValueReference() : _ptr(makeShared<T>()) {}
 
     /**
      * Instantiates a reference containing a new value of `T` initialized to
@@ -64,7 +65,7 @@ public:
      *
      * @param t value to initialize new instance with
      */
-    ValueReference(T t) : _ptr(std::make_shared<T>(std::move(t))) {}
+    ValueReference(T t) : _ptr(makeShared<T>(std::move(t))) {}
 
     /**
      * Instantiates a new reference from an existing `std::shared_ptr` to a
@@ -76,7 +77,7 @@ public:
      *
      * @param t shared pointer to link to
      */
-    explicit ValueReference(std::shared_ptr<T> t) : _ptr(std::move(t)) {}
+    explicit ValueReference(SharedPtr<T> t) : _ptr(std::move(t)) {}
 
     /**
      * Copy constructor. The new instance will refer to a copy of the
@@ -84,9 +85,9 @@ public:
      */
     ValueReference(const ValueReference& other) {
         if ( auto ptr = other._get() )
-            _ptr = std::make_shared<T>(*ptr);
+            _ptr = makeShared<T>(*ptr);
         else
-            _ptr = std::shared_ptr<T>();
+            _ptr = SharedPtr<T>();
     }
 
     /** Move constructor. */
@@ -124,10 +125,10 @@ public:
      * @throws IllegalReference if no shared pointer can be constructed for
      * the contained instance.
      */
-    std::shared_ptr<T> asSharedPtr() const {
+    SharedPtr<T> asSharedPtr() const {
         assert(_ptr.index() != std::variant_npos);
 
-        if ( auto x = std::get_if<std::shared_ptr<T>>(&_ptr) )
+        if ( auto x = std::get_if<SharedPtr<T>>(&_ptr) )
             return *x;
 
         try {
@@ -139,7 +140,7 @@ public:
             }
             else
                 throw IllegalReference("unexpected state of value reference");
-        } catch ( const std::bad_weak_ptr& ) {
+        } catch ( const BadWeakPtr& ) {
             throw IllegalReference("reference to non-heap instance");
         }
     }
@@ -148,7 +149,7 @@ public:
      * Resets the contained value to a fresh copy of a `T` value initialized
      * to its default.
      */
-    void reset() { _ptr = std::shared_ptr<T>(); }
+    void reset() { _ptr = SharedPtr<T>(); }
 
     /**
      * Returns a reference to the contained value.
@@ -264,11 +265,11 @@ public:
     }
 
     /**
-     * Assigns from an existing `std::shared_ptr` to a value of type `T`. This
+     * Assigns from an existing `SharedPtr` to a value of type `T`. This
      * does *not* copy the pointer's target value; we will store a pointer to
      * the same value.
      */
-    ValueReference& operator=(std::shared_ptr<T> other) noexcept {
+    ValueReference& operator=(SharedPtr<T> other) noexcept {
         if ( _get() != other.get() )
             _ptr = std::move(other);
 
@@ -307,16 +308,16 @@ private:
         if ( auto ptr = std::get_if<T*>(&_ptr) )
             return *ptr;
 
-        assert(std::holds_alternative<std::shared_ptr<T>>(_ptr));
-        return std::get_if<std::shared_ptr<T>>(&_ptr)->get();
+        assert(std::holds_alternative<SharedPtr<T>>(_ptr));
+        return std::get_if<SharedPtr<T>>(&_ptr)->get();
     }
 
     inline T* _get() noexcept {
         if ( auto ptr = std::get_if<T*>(&_ptr) )
             return *ptr;
 
-        assert(std::holds_alternative<std::shared_ptr<T>>(_ptr));
-        return std::get_if<std::shared_ptr<T>>(&_ptr)->get();
+        assert(std::holds_alternative<SharedPtr<T>>(_ptr));
+        return std::get_if<SharedPtr<T>>(&_ptr)->get();
     }
 
     inline const T* _safeGet() const {
@@ -326,7 +327,7 @@ private:
         if ( auto ptr = std::get_if<T*>(&_ptr) )
             return *ptr;
 
-        if ( auto ptr = std::get_if<std::shared_ptr<T>>(&_ptr); ptr && *ptr )
+        if ( auto ptr = std::get_if<SharedPtr<T>>(&_ptr); ptr && *ptr )
             return ptr->get();
 
         reference::detail::throw_null();
@@ -339,7 +340,7 @@ private:
         if ( auto ptr = std::get_if<T*>(&_ptr) )
             return *ptr;
 
-        if ( auto ptr = std::get_if<std::shared_ptr<T>>(&_ptr); ptr && *ptr )
+        if ( auto ptr = std::get_if<SharedPtr<T>>(&_ptr); ptr && *ptr )
             return ptr->get();
 
         reference::detail::throw_null();
@@ -348,7 +349,7 @@ private:
     // In `_safeGet` above we rely on the fact that a default-constructed
     // `ValueReference` always contains a `shared_ptr`, so it is listed as the
     // first variant.
-    std::variant<std::shared_ptr<T>, T*> _ptr;
+    std::variant<SharedPtr<T>, T*> _ptr;
 };
 
 /**
@@ -359,9 +360,9 @@ private:
  * explicitly be null.
  */
 template<typename T>
-class StrongReference : public std::shared_ptr<T> {
+class StrongReference : public SharedPtr<T> {
 public:
-    using Base = std::shared_ptr<T>;
+    using Base = SharedPtr<T>;
 
     /** Default constructor creating a null reference. */
     StrongReference() : Base() {}
@@ -378,7 +379,7 @@ public:
      *
      * @param t initialization value
      */
-    explicit StrongReference(T t) : Base(std::make_shared<T>(std::move(t))) {}
+    explicit StrongReference(T t) : Base(makeShared<T>(std::move(t))) {}
 
     /** Instantiate an unset reference. */
     explicit StrongReference(std::nullptr_t) {}
@@ -459,7 +460,7 @@ public:
      * @param t value to allocate and then refer to
      */
     StrongReference& operator=(T other) {
-        Base::operator=(std::make_shared<T>(std::move(other)));
+        Base::operator=(makeShared<T>(std::move(other)));
         return *this;
     }
 
@@ -511,9 +512,9 @@ private:
  * be null.
  */
 template<typename T>
-class WeakReference : public std::weak_ptr<T> {
+class WeakReference : public WeakPtr<T> {
 public:
-    using Base = std::weak_ptr<T>;
+    using Base = WeakPtr<T>;
 
     /** Default constructor creating a null reference. */
     WeakReference() = default;