--- a/hilti/runtime/src/backtrace.cc
+++ b/hilti/runtime/src/backtrace.cc
@@ -10,7 +10,7 @@ using namespace hilti::rt;
 
 Backtrace::Backtrace() {
 #ifdef HILTI_HAVE_BACKTRACE
-    _callstack = std::make_shared<Callstack>(); // allocate on heap to save stack space
+    _callstack = makeShared<Callstack>(); // allocate on heap to save stack space
     _frames = ::backtrace(_callstack->begin(), static_cast<int>(_callstack->size()));
 #endif
 }