--- a/spicy/runtime/src/base64.cc
+++ b/spicy/runtime/src/base64.cc
@@ -18,7 +18,7 @@ struct detail::State {
 };
 
 Stream::Stream() {
-    _state = std::shared_ptr<detail::State>(new detail::State(), [](auto p) {
+    _state = hilti::rt::SharedPtr<detail::State>(new detail::State(), [](auto p) {
         // Nothing else to clean up.
         delete p; // NOLINT(cppcoreguidelines-owning-memory)
     });