Closes #1600.