@awelzel 2023-11-27
The microbenchmark from #1600 doesn't show the single vs multi-threading difference anymore. This branch appears to run a tiny bit slower than `main` when single-threaded though. I haven't looked why, maybe `std::shared_ptr` is overall faster than the boost version even when atomic instructions are in-use?

```
$ hyperfine -w 1 '/opt/zeek-dev-prod/bin/spicy-driver -f test-data.txt nested-prev.hlto' 'spicy-driver -f test-data.txt nested.hlto'
Benchmark 1: /opt/zeek-dev-prod/bin/spicy-driver -f test-data.txt nested-prev.hlto
  Time (mean ± σ):     12.137 s ±  0.209 s    [User: 12.112 s, System: 0.010 s]
  Range (min … max):   11.996 s … 12.704 s    10 runs
 
Benchmark 2: spicy-driver -f test-data.txt nested.hlto
  Time (mean ± σ):     12.360 s ±  0.158 s    [User: 12.347 s, System: 0.011 s]
  Range (min … max):   12.134 s … 12.529 s    10 runs
 
Summary
  '/opt/zeek-dev-prod/bin/spicy-driver -f test-data.txt nested-prev.hlto' ran
    1.02 ± 0.02 times faster than 'spicy-driver -f test-data.txt nested.hlto'
```
`/opt/zeek-dev-prod/bin/spicy-driver` being current main.