@bbannier 2023-11-29
I benchmarked this PR against a huge real world Zeek analyzer we are developing at Corelight. In summary, it seems doubtful that this PR provides a clear benefit when running real analyzers in Zeek.

# This branch vs. `main`

Before (main):

```console
$ hyperfine -w 1 'zeek -Cr large.pcap ../build/spicy-modules/*.hlto ../analyzer/scripts'
Benchmark 1: zeek -Cr large.pcap ../build/spicy-modules/*.hlto ../analyzer/scripts
  Time (mean ± σ):     19.428 s ±  0.419 s    [User: 20.509 s, System: 1.326 s]
  Range (min … max):   18.884 s … 20.448 s    10 runs
```

After (this PR):

```console
$ hyperfine -w 1 'zeek -Cr large.pcap ../build/spicy-modules/*.hlto ../analyzer/scripts'
Benchmark 1: zeek -Cr large.pcap ../build/spicy-modules/*.hlto ../analyzer/scripts
  Time (mean ± σ):     19.223 s ±  0.371 s    [User: 20.314 s, System: 1.310 s]
  Range (min … max):   18.774 s … 19.844 s    10 runs
```

I want to see that the mean runtime of this particular benchmark seems to improve, but the difference here was not significant.

# This branch on top of #1590

Among other things in #1590 we also removed some unneeded copying, so we might expect to see different behavior of this PR since e.g., overall copying of `Bytes` instances might be reduced.

Before (#1590):

```console
$ hyperfine -w 1 'zeek -Cr large.pcap ../build/spicy-modules/*.hlto ../analyzer/scripts'
Benchmark 1: zeek -Cr large.pcap ../build/spicy-modules/*.hlto ../analyzer/scripts
  Time (mean ± σ):     17.308 s ±  0.299 s    [User: 18.322 s, System: 1.331 s]
  Range (min … max):   16.860 s … 17.813 s    10 runs
```

After (this PR on top of #1590):

```console
$ hyperfine -w 1 'zeek -Cr large.pcap ../build/spicy-modules/*.hlto ../analyzer/scripts'
Benchmark 1: zeek -Cr large.pcap ../build/spicy-modules/*.hlto ../analyzer/scripts
  Time (mean ± σ):     17.578 s ±  0.335 s    [User: 18.607 s, System: 1.297 s]
  Range (min … max):   17.120 s … 18.134 s    10 runs
```

The trend of the means seems to be reversed here, i.e., the new "improved" smart pointer implementation seems to introduce new overhead, but the result again is not significant.